requirements.txt
django~=5.0.6
django-allauth~=0.63.3
pip==24.1
attrs~=23.1.0
cryptography~=41.0.7
Pillow~=10.1.0
Jinja2~=3.1.4
filelock~=3.13.4
asgiref~=3.7.2
sqlparse~=0.4.4
requests~=2.31.0
setuptools~=69.0.3
platformdirs~=4.1.0
django-filter~=24.2
